+++ src/App.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { Routes, Route } from "react-router-dom";
 import Wiki from "./pages/WikiPage/Wiki";
 import WikiWrite from "./pages/WikiPage/WikiWrite";
-import ItemContent from "./pages/WikiPage/ItemContent";
+import WikiItem from "./pages/WikiPage/WikiItem";
 import WikiEdit from "./pages/WikiPage/WikiEdit";
 import Login from "./pages/LoginPage/Login";
 import SignUp from "./pages/LoginPage/SignUp";
@@ -33,17 +33,17 @@ function App() {
         <Route path="/notice/write" element={<NewWrite />} />
         <Route path="/notice/:noticeId/edit" element={<EditWrite />} />
         <Route path="/notice/:noticeId" element={<NoticeDetail />} />
+        <Route path="/wiki" element={<Wiki />} />
+        <Route path="/wiki/write" element={<WikiWrite />} />
+        <Route path="/wiki/content" element={<WikiItem />} />
+        <Route path="/wiki/edit" element={<WikiEdit />} />
       </Route>
       <Route element={<PublicRoute />}>
         <Route path="/project/:projectId" element={<ProjectDetail />} />
         <Route path="/projectwrite" element={<ProjectWrite />} />
         <Route path="/project/edit/:projectId" element={<ProjectEdit />} />
         <Route path="/projectlist" element={<ProjectList />} />
-        <Route path="/" element={<Main />} />
-        <Route path="/wiki" element={<Wiki />} />
-        <Route path="/wiki/write" element={<WikiWrite />} />
-        <Route path="/wiki/content" element={<ItemContent />} />
-        <Route path="/wiki/edit" element={<WikiEdit />} />
+        
       </Route>
     </Routes>
   );
